api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
gunicorn==21.2.0
